PluginManager and hooking it up to ClientManager.
Events are now done as an event object (I was worried about introducing
new parameters later and this allows some insulation from that.)
ClientManager now supports events as well and will relay any events
from itself with an additional event parameter indicating the client it
came from.
Plugin names are found in the metadata hash for the plugin and
currently have to match the name of the file to require to load that
plugin. I realise that this is less than ideal.